@zhassan-aws 2024-03-11
I ran the main test that has failed the performance regressions (`s2n-quic/quic/s2n-quic-core/sync::spsc::tests::alloc_test`) 3 times locally with and without the changes, and while there is some slowdown, it's within an acceptable percentage (~9%):
With the toolchain upgrade:
```
Verification Time: 70.934296s
Verification Time: 70.673645s
Verification Time: 70.516945s
```
Without the toolchain upgrade:
```
Verification Time: 64.85171s
Verification Time: 64.96482s
Verification Time: 65.115456s
```